dex_arb/requirements.txt
web3==6.19.0
eth-abi==4.2.1
eth-utils==4.1.0
requests==2.32.3
PyYAML==6.0.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
loguru==0.7.2
prometheus-client==0.19.0
redis==5.0.1
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
